backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
# AI & Document Processing Libraries
PyMuPDF
python-docx
beautifulsoup4
lxml
pdfplumber